add error messages test
