--- skyVegasTests/errorDisplayedGameClient.spec.ts
@@ -1,40 +1,0 @@
-import { test, type Page } from "@playwright/test";
-import { launchGame, validateConsoleMessages, login, startEventListener, makeSpin, readGames } from "./utilities";
-import { ExpectedMessage } from "../types/expectedMessage";
-
-
-
-readGames().forEach((game: any) => {
-    let page: Page;
-    const consoleMessages: string[] = [];
-    test.describe(`Testing with text: ${game}`, () => {
-      test.beforeAll(async ({ browser }) => {
-        page = await browser.newPage();
-        await login(page);
-      });
-  
-      test.afterAll(async () => {
-        await page.close();
-      });
-  
-      test("errorDisplayed", async () => {
-        startEventListener(page, consoleMessages);
-
-    await launchGame(page, game, consoleMessages);
-
-    await makeSpin(page, consoleMessages);
-
-    // Wait 60s so the reality check timer kicks in
-    await page.waitForTimeout(60000)
-
-    await makeSpin(page, consoleMessages);
-
-    await validateConsoleMessages(
-        ExpectedMessage.ERROR_DISPLAYED,
-        consoleMessages,
-        [65_000, 70_000, 75_000],
-        120_000
-    )
-      });
-    });
-  });
